fix: moved tx from confirm
