+++ src/renderer/features/multisig-wallet-create/flexible-multisig/model/flexible-multisig-create.ts
@@ -410,6 +410,7 @@ export const flexibleMultisigModel = {
   $pendingFee: or($pendingProxyFee, $pendingMultisigFee),
   $proxyDeposit,
   $existentialDeposit,
+  $totalDeposit,
   $isLoading: or($pendingProxyFee, $pendingMultisigFee, getExistentialDepositFx.pending),
   $isEnoughBalance,
